add balance the db
