+++ app/components/AddBalance.tsx
@@ -2,29 +2,14 @@
 
 import { useAtom } from 'jotai';
 import React, { FormEvent, useEffect, useState } from 'react';
-import { balanceAtom, userAtom } from '../store/atoms';
+import { userAtom } from '../store/atoms';
 import { createTransactionDocument, db } from '@/firebase/firestore';
-import { doc, getDoc } from 'firebase/firestore';
+import { useGetBalance } from '../hooks/useGetBalance';
 
 const AddBalance = () => {
   const [user] = useAtom(userAtom);
   const [input, setInput] = useState(0);
-  const [, setBalance] = useAtom(balanceAtom);
-
-  useEffect(() => {
-    const getBalance = async () => {
-      const docRef = doc(db, 'transactions', user.uid);
-      const docSnap = await getDoc(docRef);
-      if (docSnap.exists()) {
-        const docData = docSnap.data();
-        const balance = docData.balance;
-        setBalance(balance);
-      } else {
-        console.log('No such document!');
-      }
-    };
-    getBalance();
-  }, []);
+  useGetBalance();
 
   const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
